input: goodix: force enable hover proximity sensor
The sensors HAL passes "1" to the ear_detect_enable sec_cmd, which leads to the proximity sensor being enabled.

So far it seems like 1 enables touch proximity, whereas 3 enables both hover proximity and touch proximity. It is yet to figure out if and what 2 does, but for now force enable the hover feature by altering the value passed by the HAL to 3.

Only enable hovering when the screen is not in LPM because turning on the hover feature breaks proximity when in LPM. While in LPM touch proximity will be used, otherwise touch proximity will be disabled.

Change-Id: I7d77a79772d4661c5d1dc7dc64e144299c8f26d5
Signed-off-by: fiqys <elsifycritic@gmail.com>